evo Better error management
Change the way errors are handled to have the same way in each view
Adding custom Http error
More factoring

+++ projets/views/note_view.py
@@ -1,8 +1,16 @@
 from django.contrib.auth.decorators import login_required
 from django.utils.translation import gettext
 
-from projets.commun.error import HttpStatus, render_error, raise_suspicious_operation
-from projets.commun.utils import java_string_hashcode, JsonResponseContainer
+from projets.commun.error import (
+    HttpInvalidFormData,
+    HttpForbidden,
+    HttpMethodNotAllowed,
+)
+from projets.commun.utils import (
+    java_string_hashcode,
+    JsonResponseContainer,
+    get_object_or_not_found,
+)
 from projets.forms import NoteForm
 from projets.models import Note, get_current_timestamp, Codex, Page
 
@@ -48,31 +56,24 @@ def post_note(request):
 
     # If the form is not valid, throw an error
     if not input_form.is_valid():
-        # TODO: review the form validation
-        form_errors = gettext("Form validation error.")
-        response.data.update({"success": False, "form_errors": form_errors})
-        response.status = 400
-        return response.get_json_response()
+        raise HttpInvalidFormData(
+            form_errors=input_form.non_field_errors(),
+            fields_error=input_form.errors.as_json(),
+        )
 
     # Get the codex of from the slug
     codex_slug = request.POST.get("codex_slug")
 
     # If the codex_slug is empty, throw an error
     if not codex_slug:
+        # TODO : as the codex slug is in the POST param, it should be in the form validation
         form_errors = gettext("Form validation error.")
         response.data.update({"success": False, "form_errors": form_errors})
         response.status = 400
         return response.get_json_response()
 
     # Get the codex of the note from the slug
-    try:
-        codex = Codex.objects.get(slug=codex_slug)
-    except Codex.DoesNotExist:
-        # TODO : factorise this
-        local_error = gettext("The Codex does not exist.")
-        response.data.update({"success": False, "local_error": local_error})
-        response.status = 404
-        return response.get_json_response()
+    codex = get_object_or_not_found(Codex, slug=codex_slug)
 
     # Check if a note_details_view already exist for today
     # TODO: try to upgrade the way this check is down
@@ -81,20 +82,17 @@ def post_note(request):
 
     # If a note_details_view was already created today, throw an error
     if hasattr(page, "note"):
-        local_error = gettext("A Note already exist for today.")
-        response.data.update({"success": False, "local_error": local_error})
-        response.status = 400
-        return response.get_json_response()
+        raise HttpInvalidFormData(
+            form_errors=gettext("A Note already exist for today for this Codex.")
+        )
 
     # Check if the user has the permission
     if not is_authorized_to_create_note(request.user, codex):
-        local_error = gettext("You are not allowed to create this Note")
-        response.data.update({"success": False, "local_error": local_error})
-        response.status = 403
-        return response.get_json_response()
+        raise HttpForbidden
 
     # If all the above test are ok, create the note_details_view
     request_text = input_form.save(commit=False).text
+    # TODO : use the form to save the note
     note = Note(text=request_text)
     note.save(codex=codex)
 
@@ -118,48 +116,33 @@ def put_note(request, note_id):
 
     # If the form is not valid, throw an error
     if not input_form.is_valid():
-        # TODO : review the validation of the form
-        form_errors = gettext("Form validation error.")
-        response.data.update(
-            {"success": False, "id": note_id, "form_errors": form_errors}
+        raise HttpInvalidFormData(
+            form_errors=input_form.non_field_errors(),
+            fields_error=input_form.errors.as_json(),
         )
-        response.status = 400
-        return response.get_json_response()
 
     # Get the text from the form
     request_text = input_form.save(commit=False).text
 
-    # Get the note_details_view from the database
-    try:
-        note = Note.objects.get(id=note_id)
-    except Note.DoesNotExist:
-        local_error = gettext("The Note does not exist.")
-        response.data.update({"success": False, "local_error": local_error})
-        response.status = 404
-        return response.get_json_response()
+    # Get the note from the database
+    note = get_object_or_not_found(Note, id=note_id)
 
     # If the note_details_view hash of the request doesn't correspond to the note_details_view hash of the database,
     #  throw an error
     database_note_hash = str(java_string_hashcode(note.text))
     if request_note_hash != database_note_hash:
-        local_error = gettext(
-            "The Note have been modified since the last modification attempt."
+        raise HttpInvalidFormData(
+            form_errors=gettext(
+                "The Note have been modified since the last modification attempt."
+            )
         )
-        response.data.update(
-            {"success": False, "id": note_id, "local_error": local_error}
-        )
-        response.status = 400
-        return response.get_json_response()
 
     # Check if the user has the permission
     if not is_authorized_to_update_note(request.user, note):
-        # TODO : factorise this
-        local_error = gettext("You are note authorized to modify this Note.")
-        response.data.update({"success": False, "local_error": local_error})
-        response.status = 403
-        return response.get_json_response()
+        raise HttpForbidden
 
     # If all the above test are ok, update the note_details_view
+    # TODO : use the form to save the note
     note.text = request_text
     note.save()
 
@@ -177,14 +160,15 @@ def delete_note(request, note_id):
 
     # Check if the user has the permission
     if not is_authorized_to_delete_note(request.user, note_id):
-        # TODO : factorise this
-        local_error = gettext("You are note authorized to delete this Note.")
-        response.data.update({"success": False, "local_error": local_error})
-        response.status = 403
-        return response.get_json_response()
+        raise HttpForbidden
 
-    # Delete the note_details_view
-    Note.objects.get(id=note_id).delete()
+    # Get the note from the database
+    note = get_object_or_not_found(Note, id=note_id)
+
+    # Delete the note
+    # TODO : add a test if the not does not exist
+    # TODO : check if other delete use the same pattern (get the Model THEN delete it)
+    note.delete()
 
     # Prepare the output data
     response.data.update({"success": True, "id": note_id})
@@ -196,26 +180,15 @@ def note_details_view(request, note_id):
     """
     Manage REST like actions on a note.
     """
-    # Create the output data
-    http_status = HttpStatus()
-    response = None
+    if request.method == "PUT" and request.is_ajax():
+        response = put_note(request, note_id)
+    elif request.method == "DELETE" and request.is_ajax():
+        response = delete_note(request, note_id)
+    else:
+        raise HttpMethodNotAllowed
 
-    try:
-        if not request.is_ajax():
-            raise_suspicious_operation(http_status)
-
-        if request.method == "PUT":
-            response = put_note(request, note_id)
-        elif request.method == "DELETE":
-            response = delete_note(request, note_id)
-        else:
-            raise_suspicious_operation(http_status)
-
-        # Return the output data
-        return response
-    except Exception as ex:
-        # Return the error as a html page or as a JSON dictionary
-        return render_error(request, ex, http_status)
+    # Return the output data
+    return response
 
 
 @login_required
@@ -223,21 +196,10 @@ def note_list_view(request):
     """
     Manage REST like actions on notes.
     """
-    # Create the output data
-    http_status = HttpStatus()
-    response = None
+    if request.method == "POST" and request.is_ajax():
+        response = post_note(request)
+    else:
+        raise HttpMethodNotAllowed
 
-    try:
-        if not request.is_ajax():
-            raise_suspicious_operation(http_status)
-
-        if request.method == "POST":
-            response = post_note(request)
-        else:
-            raise_suspicious_operation(http_status)
-
-        # Return the output data
-        return response
-    except Exception as ex:
-        # Return the error as a html page or as a JSON dictionary
-        return render_error(request, ex, http_status)
+    # Return the output data
+    return response
